Refactor 1: Simplify property derivation in CExpression.
ORCA recursively derives properties of an operator's children before it
calls Derive() on the top operator. Before this, the recursion was
implemented in a co-routine style, with the calls bouncing between
CDrvdPropXXX (via CExpression) & CExpressionHandle. Furthermore, the derived
properties are saved in both CExpression & CExpressionHandle supposedly
for easier access.

This commits tries to simplify the property derivation logic, so that we
can later exploit on-demand property derivation. It also updates the
getters in CExpressionHandle to retrieve the properties from the
underlying CExpression (or cached properties in CGroupExpression)
instead of using it's own separate cache.

Note that this was not done all the way. CExpressionHandle still caches
pdpplan for some corner cases, we weren't able to refactor around at
this time.

This commit also removes code we no longer need.